requirements.txt
streamlit==1.43.2
python-dotenv==1.0.1
langchain==0.3.21
langchain-core==0.3.47
langchain-community==0.3.20
langchain-google-genai==2.1.0
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
openai==1.68.0
pydantic==2.10.6
pydantic-settings==2.8.1
protobuf==5.29.4
numpy==2.2.4
pandas==2.2.3
tiktoken==0.9.0
fastapi==0.115.11
langserve==0.3.1
uvicorn==0.34.0